@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -33,6 +33,9 @@ dependencies:
   barcode_scan: any
   http: ^0.12.0+2
   sqflite: ^1.1.6
+  path_provider: ^1.3.0
+  mobx: ^0.3.7
+  flutter_mobx: 0.3.0+1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -41,6 +44,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  build_runner: ^1.3.1
+  mobx_codegen: 0.3.3+1
 
 
 # For information on the generic Dart part of this file, see the
