@@ -1,3 +1,4 @@
+import 'package:cake_wallet/src/domain/common/wallet_description.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:cake_wallet/src/domain/common/wallet.dart';
 import 'package:cake_wallet/src/domain/common/transaction_history.dart';
@@ -7,34 +8,42 @@ import 'package:cake_wallet/src/domain/common/transaction_creation_credentials.d
 import 'package:cake_wallet/src/domain/common/pending_transaction.dart';
 
 class WalletService extends Wallet {
-  Function(Wallet wallet) onWalletChange;
-  
+  get onWalletChange => _onWalletChanged.stream;
+
   get onBalanceChange => _onBalanceChange.stream;
+
   get syncStatus => _syncStatus.stream;
+
   get syncStatusValue => _syncStatus.value;
+
   get walletType => _currentWallet.walletType;
+
   get currentWallet => _currentWallet;
 
   set currentWallet(Wallet wallet) {
     _currentWallet = wallet;
     _currentWallet.onBalanceChange.listen((wallet) => _onBalanceChange.add(wallet));
     _currentWallet.syncStatus.listen((status) => _syncStatus.add(status));
+    _onWalletChanged.add(wallet);
 
-    if (onWalletChange != null) {
-      onWalletChange(wallet);
-    }
+    final type = wallet.getType();
+    wallet.getName().then((name) => description = WalletDescription(name: name, type: type));
   }
 
+  BehaviorSubject<Wallet> _onWalletChanged;
   BehaviorSubject<Wallet> _onBalanceChange;
   BehaviorSubject<SyncStatus> _syncStatus;
   Wallet _currentWallet;
 
-  WalletService({this.onWalletChange}) {
+  WalletService() {
     _currentWallet = null;
     walletType = WalletType.NONE;
     _syncStatus = BehaviorSubject<SyncStatus>();
     _onBalanceChange = BehaviorSubject<Wallet>();
+    _onWalletChanged = BehaviorSubject<Wallet>();
   }
+
+  WalletDescription description;
 
   WalletType getType() => WalletType.MONERO;
 
@@ -56,24 +65,17 @@ class WalletService extends Wallet {
 
   Future<bool> isConnected() => _currentWallet.isConnected();
 
+  Future<void> close() => _currentWallet.close();
+
   Future<void> connectToNode(
-          {String uri,
-          String login,
-          String password,
-          bool useSSL = false,
-          bool isLightWallet = false}) =>
+          {String uri, String login, String password, bool useSSL = false, bool isLightWallet = false}) =>
       _currentWallet.connectToNode(
-          uri: uri,
-          login: login,
-          password: password,
-          useSSL: useSSL,
-          isLightWallet: isLightWallet);
+          uri: uri, login: login, password: password, useSSL: useSSL, isLightWallet: isLightWallet);
 
   Future<void> startSync() => _currentWallet.startSync();
 
   TransactionHistory getHistory() => _currentWallet.getHistory();
 
-  Future<PendingTransaction> createTransaction(
-          TransactionCreationCredentials credentials) =>
+  Future<PendingTransaction> createTransaction(TransactionCreationCredentials credentials) =>
       _currentWallet.createTransaction(credentials);
 }
