@@ -8,105 +8,51 @@ import 'package:cake_wallet/src/screens/dashboard/sync_info.dart';
 import 'package:cake_wallet/src/screens/receive/qr_image.dart';
 
 class Receive extends StatelessWidget {
+  final _closeButtonImage = Image.asset('assets/images/close_button.png');
+  final _shareButtonImage = Image.asset('assets/images/share_button.png');
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         resizeToAvoidBottomPadding: false,
         backgroundColor: Colors.white,
-        body: SafeArea(
-            child: Consumer<WalletInfo>(builder: (_, walletInfo, child) {
-          return ReceiveFrom(walletInfo.subaddress.address, {});
-        })));
+        appBar: CupertinoNavigationBar(
+          leading: ButtonTheme(
+            minWidth: double.minPositive,
+            child: FlatButton(
+                onPressed: () => Navigator.of(context).pop(),
+                child: _closeButtonImage),
+          ),
+          middle: Text(
+            'Receive',
+            style: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w600),
+          ),
+          trailing: ButtonTheme(
+            minWidth: double.minPositive,
+            child: FlatButton(onPressed: () {}, child: _shareButtonImage),
+          ),
+          backgroundColor: Colors.white,
+          border: null,
+        ),
+        body: SafeArea(child: SingleChildScrollView(child: ReceiveBody())));
   }
 }
 
-class ReceiveFrom extends StatefulWidget {
-  final String address;
-  final Map<String, String> subaddressMap;
-
-  const ReceiveFrom(this.address, this.subaddressMap);
-
-  @override
-  _ReceiveStateFrom createState() => _ReceiveStateFrom();
-}
-
-class _ReceiveStateFrom extends State<ReceiveFrom> {
-  final _closeButtonImage = Image.asset('assets/images/close_button.png');
-  final _shareButtonImage = Image.asset('assets/images/share_button.png');
-
-  String _qrText;
-
-  @override
-  void initState() {
-    super.initState();
-    _qrText = 'monero:' + widget.address;
-  }
-
-  void _validateAmount(String amount) {
-    String p = '^[0-9]{1,10}([.][0-9]{0,10})?\$';
-    RegExp regExp = new RegExp(p);
-    _qrText = 'monero:' + widget.address;
-    if (regExp.hasMatch(amount)) {
-      _qrText += '?tx_amount=' + amount;
-    }
-    setState(() {});
-  }
-
+class ReceiveBody extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final walletProvider = Provider.of<WalletInfo>(context);
 
     return SafeArea(
         child: SingleChildScrollView(
-      child: Column(
-        children: <Widget>[
-          Container(
-            height: 37.0,
-            margin: EdgeInsets.only(
-              top: 5.0,
-            ),
-            padding: EdgeInsets.only(left: 18.0, right: 18.0),
-            child: Stack(
-              children: <Widget>[
-                Positioned(
-                    left: 0.0,
-                    child: Container(
-                      width: 37.0,
-                      height: 37.0,
-                      child: InkWell(
-                        onTap: () {
-                          Navigator.of(context).pop();
-                        },
-                        child: _closeButtonImage,
-                      ),
-                    )),
-                Container(
-                  height: 37.0,
-                  alignment: Alignment.center,
-                  child: Text(
-                    'Receive',
-                    style:
-                        TextStyle(fontSize: 16.0, fontWeight: FontWeight.w600),
-                  ),
-                ),
-                Positioned(
-                    right: 10.0,
-                    child: Container(
-                      width: 37.0,
-                      height: 37.0,
-                      child: InkWell(
-                        onTap: () {},
-                        child: _shareButtonImage,
-                      ),
-                    ))
-              ],
-            ),
-          ),
-          Container(
-            padding: EdgeInsets.all(35.0),
             child: Column(
-              children: <Widget>[
-                Row(
+      children: <Widget>[
+        Container(
+          padding: EdgeInsets.all(35.0),
+          child: Column(
+            children: <Widget>[
+              Consumer<WalletInfo>(builder: (context, walletInfo, child) {
+                return Row(
                   children: <Widget>[
                     Spacer(
                       flex: 1,
@@ -116,7 +62,7 @@ class _ReceiveStateFrom extends State<ReceiveFrom> {
                         child: AspectRatio(
                           aspectRatio: 1.0,
                           child: QrImage(
-                            data: _qrText,
+                            data: walletInfo.subaddress.address,
                             backgroundColor: Colors.white,
                           ),
                         )),
@@ -124,8 +70,10 @@ class _ReceiveStateFrom extends State<ReceiveFrom> {
                       flex: 1,
                     )
                   ],
-                ),
-                Row(
+                );
+              }),
+              Consumer<WalletInfo>(builder: (context, walletInfo, child) {
+                return Row(
                   children: <Widget>[
                     Expanded(
                         child: Container(
@@ -133,18 +81,19 @@ class _ReceiveStateFrom extends State<ReceiveFrom> {
                       child: Center(
                         child: GestureDetector(
                           onTap: () {
-                            Clipboard.setData(
-                                ClipboardData(text: widget.address));
-                            Scaffold.of(context).showSnackBar(
-                              SnackBar(
-                                content: Text('Copied to Clipboard'),
-                                backgroundColor: Colors.green,
-                                duration: Duration(milliseconds: 1500),
+                            Clipboard.setData(ClipboardData(
+                                text: walletInfo.subaddress.address));
+                            Scaffold.of(context).showSnackBar(SnackBar(
+                              content: Text(
+                                'Copied to Clipboard',
+                                textAlign: TextAlign.center,
+                                style: TextStyle(color: Colors.black),
                               ),
-                            );
+                              backgroundColor: Palette.purple,
+                            ));
                           },
                           child: Text(
-                            widget.address,
+                            walletInfo.subaddress.address,
                             textAlign: TextAlign.center,
                             style: TextStyle(
                                 fontSize: 14.0, fontWeight: FontWeight.w600),
@@ -153,113 +102,106 @@ class _ReceiveStateFrom extends State<ReceiveFrom> {
                       ),
                     ))
                   ],
-                ),
-                Row(
-                  children: <Widget>[
-                    Expanded(
-                        child: TextField(
-                      keyboardType:
-                          TextInputType.numberWithOptions(decimal: true),
-                      decoration: InputDecoration(
-                          hintStyle: TextStyle(color: Palette.lightBlue),
-                          hintText: 'Amount',
-                          focusedBorder: UnderlineInputBorder(
-                              borderSide: BorderSide(
-                                  color: Palette.lightGrey, width: 2.0)),
-                          enabledBorder: UnderlineInputBorder(
-                              borderSide: BorderSide(
-                                  color: Palette.lightGrey, width: 2.0))),
-                      onSubmitted: (value) {
-                        _validateAmount(value);
-                      },
-                    ))
-                  ],
-                )
-              ],
-            ),
+                );
+              }),
+              Row(
+                children: <Widget>[
+                  Expanded(
+                      child: TextField(
+                    keyboardType:
+                        TextInputType.numberWithOptions(decimal: true),
+                    decoration: InputDecoration(
+                        hintStyle: TextStyle(color: Palette.lightBlue),
+                        hintText: 'Amount',
+                        focusedBorder: UnderlineInputBorder(
+                            borderSide: BorderSide(
+                                color: Palette.lightGrey, width: 2.0)),
+                        enabledBorder: UnderlineInputBorder(
+                            borderSide: BorderSide(
+                                color: Palette.lightGrey, width: 2.0))),
+                    onSubmitted: (value) {
+                      // _validateAmount(value);
+                    },
+                  ))
+                ],
+              )
+            ],
           ),
-          Row(
-            children: <Widget>[
-              Expanded(
-                  child: Container(
-                color: Palette.lightGrey2,
-                child: Column(
-                  children: <Widget>[
-                    ListTile(
-                      title: Text(
-                        'Subaddresses',
-                        style: TextStyle(fontSize: 16.0),
-                      ),
-                      trailing: Container(
-                        width: 28.0,
-                        height: 28.0,
-                        decoration: BoxDecoration(
-                            color: Color.fromRGBO(
-                                131, 87, 255, 0.2), // FIXME: Hardcoded value
-                            shape: BoxShape.circle),
-                        child: InkWell(
-                          onTap: () => Navigator.of(context)
-                              .pushNamed(newSubaddressRoute),
-                          borderRadius: BorderRadius.all(Radius.circular(14.0)),
-                          child: Icon(
-                            Icons.add,
-                            color: Palette.violet,
-                            size: 20.0,
-                          ),
+        ),
+        Row(
+          children: <Widget>[
+            Expanded(
+                child: Container(
+              color: Palette.lightGrey2,
+              child: Column(
+                children: <Widget>[
+                  ListTile(
+                    title: Text(
+                      'Subaddresses',
+                      style: TextStyle(fontSize: 16.0),
+                    ),
+                    trailing: Container(
+                      width: 28.0,
+                      height: 28.0,
+                      decoration: BoxDecoration(
+                          color: Palette.purple, shape: BoxShape.circle),
+                      child: InkWell(
+                        onTap: () => Navigator.of(context).pushNamed(newSubaddressRoute),
+                        borderRadius: BorderRadius.all(Radius.circular(14.0)),
+                        child: Icon(
+                          Icons.add,
+                          color: Palette.violet,
+                          size: 20.0,
                         ),
                       ),
                     ),
-                    Divider(
-                      color: Palette.lightGrey,
-                      height: 1.0,
-                    )
-                  ],
-                ),
-              ))
-            ],
-          ),
-          Expanded(
-              child: Container(
-                  padding: EdgeInsets.only(bottom: 20.0),
-                  child: Consumer<SubaddressListInfo>(
-                      builder: (context, subaddressInfo, snapshot) {
-                    return ListView.separated(
-                        physics: NeverScrollableScrollPhysics(),
-                        separatorBuilder: (BuildContext context, int index) =>
-                            Divider(
-                              color: Palette.lightGrey,
-                              height: 1.0,
-                            ),
-                        itemCount: subaddressInfo.subaddresses.length,
-                        itemBuilder: (BuildContext context, int index) {
-                          final subaddress = subaddressInfo.subaddresses[index];
-                          var label = subaddress.label;
+                  ),
+                  Divider(
+                    color: Palette.lightGrey,
+                    height: 1.0,
+                  )
+                ],
+              ),
+            ))
+          ],
+        ),
+        Consumer<SubaddressListInfo>(builder: (context, subaddressInfo, child) {
+          return ListView.separated(
+              shrinkWrap: true,
+              physics: NeverScrollableScrollPhysics(),
+              itemCount: subaddressInfo.subaddresses.length,
+              separatorBuilder: (context, i) {
+                return Divider(
+                  color: Palette.lightGrey,
+                  height: 1.0,
+                );
+              },
+              itemBuilder: (context, i) {
+                final subaddress = subaddressInfo.subaddresses[i];
+                final isCurrent =
+                    walletProvider.subaddress.address == subaddress.address;
+                final label = subaddress.label.isNotEmpty
+                    ? subaddress.label
+                    : subaddress.address;
 
-                          if (label == null || label.isEmpty) {
-                            label = subaddress.address;
-                          }
-
-                          return Container(
-                            color: subaddress.id == walletProvider.subaddress.id
-                                ? Palette.purple
-                                : Palette.lightGrey2,
-                            child: Column(
-                              children: <Widget>[
-                                ListTile(
-                                  title: Text(
-                                    label,
-                                    style: TextStyle(fontSize: 14.0),
-                                  ),
-                                  onTap: () => walletProvider
-                                      .changeCurrentSubaddress(subaddress),
-                                ),
-                              ],
-                            ),
-                          );
-                        });
-                  })))
-        ],
-      ),
-    ));
+                return Container(
+                  color: isCurrent ? Palette.purple : Palette.lightGrey2,
+                  child: Column(
+                    children: <Widget>[
+                      ListTile(
+                        title: Text(
+                          label,
+                          style: TextStyle(fontSize: 16.0),
+                        ),
+                        onTap: () =>
+                            walletProvider.changeCurrentSubaddress(subaddress),
+                      ),
+                    ],
+                  ),
+                );
+              });
+        })
+      ],
+    )));
   }
 }
