@@ -1,110 +1,77 @@
+import 'package:cake_wallet/src/screens/dashboard/sync_info.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
+import 'package:provider/provider.dart';
 import 'package:cake_wallet/palette.dart';
 import 'package:cake_wallet/src/widgets/primary_button.dart';
 
-class Seed extends StatefulWidget {
-
-  @override
-  createState() => _SeedState();
-
-}
-
-class _SeedState extends State<Seed>{
-  final _formKey = GlobalKey<FormState>();
+class Seed extends StatelessWidget {
+  String _seed;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.white,
-      resizeToAvoidBottomPadding: false,
-      appBar: CupertinoNavigationBar(
-          middle: Text('Seed', style: TextStyle(fontWeight: FontWeight.normal),),
-          backgroundColor: Colors.white,
-          border: null,
-          trailing: InkWell(
-            onTap: (){},
-            child: Text('Done',
-              style: TextStyle(
-                  fontWeight: FontWeight.bold,
-                  fontSize: 18.0
-              ),
+        backgroundColor: Colors.white,
+        resizeToAvoidBottomPadding: false,
+        appBar: CupertinoNavigationBar(
+            middle: Text(
+              'Seed',
+              style: TextStyle(fontWeight: FontWeight.normal),
             ),
-          )
-      ),
-      body: Form(
-        key: _formKey,
-        child: GestureDetector(
-          onTap: (){
+            backgroundColor: Colors.white,
+            border: null,
+            leading: Container(),
+            trailing: FlatButton(
+              padding: EdgeInsets.all(0),
+              onPressed: () => Navigator.of(context).pop(),
+              child: Text(
+                'Done',
+                style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18.0),
+              ),
+            )),
+        body: Form(
+            child: GestureDetector(
+          onTap: () {
             SystemChannels.textInput.invokeMethod('TextInput.hide');
           },
           child: Center(
             child: Container(
-              padding: EdgeInsets.only(
-                  left: 30.0,
-                  right: 30.0
-              ),
+              padding: EdgeInsets.only(left: 30.0, right: 30.0),
               child: Column(
                 mainAxisSize: MainAxisSize.min,
                 children: <Widget>[
                   Container(
-                    margin: EdgeInsets.only(
-                        left: 30.0,
-                        top: 30.0,
-                        right: 30.0,
-                        bottom: 30.0
-                    ),
-                    child: Column(
-                      children: <Widget>[
-                        Row(
-                          children: <Widget>[
-                            Expanded(
-                                child: Container(
-                                  padding: EdgeInsets.only(
-                                      bottom: 20.0
-                                  ),
-                                  child: TextFormField(
-                                    decoration: InputDecoration(
-                                        focusedBorder: UnderlineInputBorder(
-                                            borderSide: BorderSide(
-                                                color: Palette.lightGrey,
-                                                width: 1.0
-                                            )
-                                        ),
-                                        enabledBorder: UnderlineInputBorder(
-                                            borderSide: BorderSide(
-                                                color: Palette.lightGrey,
-                                                width: 1.0
-                                            )
-                                        )
-                                    ),
-                                    initialValue: 'MyAwesomeWallet',
-                                    textAlign: TextAlign.center,
-                                    validator: (value){
-                                      if (value.isEmpty) return 'Please write a seed';
-                                      return null;
-                                    },
-                                  ),
-                                )
-                            )
-                          ],
-                        ),
-                        Text(
-                          'umbrella hire boat adopt sieve money business '
-                          'royal zones repent inflamed eavesdrop cube '
-                          'umpire javelin pulp dash bikini major aloof '
-                          'hippo shocking lopped merger money',
-                          textAlign: TextAlign.center,
-                        )
-                      ],
-                    ),
+                    margin: EdgeInsets.only(left: 30.0, top: 30.0, right: 30.0, bottom: 30.0),
+                    child: Consumer<WalletSeedInfo>(builder: (context, seedInfo, child) {
+                      _seed = seedInfo.seed;
+                      return Column(
+                        children: <Widget>[
+                          Row(
+                            children: <Widget>[
+                              Expanded(
+                                  child: Container(
+                                decoration: BoxDecoration(
+                                    border: Border(bottom: BorderSide(width: 1.0, color: Palette.lightGrey))),
+                                padding: EdgeInsets.only(bottom: 20.0),
+                                margin: EdgeInsets.only(bottom: 10.0),
+                                child: Text(
+                                  seedInfo.name,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ))
+                            ],
+                          ),
+                          Text(
+                            seedInfo.seed,
+                            textAlign: TextAlign.center,
+                          )
+                        ],
+                      );
+                    }),
                   ),
                   Container(
-                    margin: EdgeInsets.only(
-                      top: 30.0,
-                      bottom: 30.0
-                    ),
+                    margin: EdgeInsets.only(top: 30.0, bottom: 30.0),
                     child: Column(
                       children: <Widget>[
                         Text(
@@ -117,35 +84,34 @@ class _SeedState extends State<Seed>{
                           ),
                         ),
                         Container(
-                          padding: EdgeInsets.only(
-                            top: 20.0
-                          ),
+                          padding: EdgeInsets.only(top: 20.0),
                           child: Row(
                             children: <Widget>[
                               Flexible(
-                                child: Container(
-                                  padding: EdgeInsets.only(
-                                    right: 15.0
-                                  ),
-                                  child: PrimaryButton(
-                                    onPressed: (){},
+                                  child: Container(
+                                padding: EdgeInsets.only(right: 15.0),
+                                child: PrimaryButton(
+                                    onPressed: () {},
                                     color: Palette.indigo,
                                     borderColor: Palette.deepIndigo,
-                                    text: 'Save'
-                                  ),
-                                )
-                              ),
+                                    text: 'Save'),
+                              )),
                               Flexible(
-                                child: Container(
-                                  padding: EdgeInsets.only(
-                                    left: 15.0
-                                  ),
-                                  child: PrimaryButton(
-                                    onPressed: (){},
-                                    text: 'Copy'
-                                  ),
-                                )
-                              )
+                                  child: Container(
+                                padding: EdgeInsets.only(left: 15.0),
+                                child: PrimaryButton(
+                                    onPressed: () {
+                                      Clipboard.setData(ClipboardData(text: _seed));
+                                      Scaffold.of(context).showSnackBar(
+                                        SnackBar(
+                                          content: Text('Copied to Clipboard'),
+                                          backgroundColor: Colors.green,
+                                          duration: Duration(milliseconds: 1500),
+                                        ),
+                                      );
+                                    },
+                                    text: 'Copy'),
+                              ))
                             ],
                           ),
                         )
@@ -156,9 +122,6 @@ class _SeedState extends State<Seed>{
               ),
             ),
           ),
-        )
-      )
-    );
+        )));
   }
-
 }
